requirements.txt
nltk==3.8.1
numpy==1.23.5
openai==0.27.4
python-dotenv==1.0.0
scikit_learn==1.2.2
tiktoken==0.3.1
